+++ contracts/interfaces/insta/IInstaIndex.sol
@@ -2,14 +2,9 @@
 pragma solidity =0.8.24;
 
 interface IInstaIndex {
-
     event LogAccountCreated(address sender, address indexed owner, address indexed account, address indexed origin);
 
-    function build(
-        address _owner,
-        uint accountVersion,
-        address _origin
-    ) external returns (address _account);
+    function build(address _owner, uint256 accountVersion, address _origin) external returns (address _account);
 
     function master() external view returns (address);
 }
